non-functional lapiz conversion
