+++ include/Networking/NetworkPlayerEntitlementChecker.hpp
@@ -1,0 +1,6 @@
+#include "custom-types/shared/macros.hpp"
+
+DECLARE_CLASS_CODEGEN(MultiplayerCore::Networking, NetworkPlayerEntitlementChecker,
+    
+
+)
