+++ include/main.hpp
@@ -11,10 +11,24 @@
 #include "beatsaber-hook/shared/utils/il2cpp-utils.hpp"
 #include "custom-types/shared/delegate.hpp"
 
+// Lapiz is an abstraction layer between Zenject and the user, making their lives easier
+#include "lapiz/shared/zenject/Zenjector.hpp"
+#include "lapiz/shared/zenject/Location.hpp"
+
+// Zenject includes
+#include "Zenject/DiContainer.hpp"
+
+// MpCore Installers
+#include "Installers/AppInstaller.hpp"
+#include "Installers/GameInstaller.hpp"
+#include "Installers/MenuInstaller.hpp"
+
 // Cpp includes
 #include <map>
 #include <set>
 #include <string>
+
+#include "logging.hpp"
 
 // Define these functions here so that we can easily read configuration and log information from other files
 Configuration& getConfig();
